Mental_health_chatbot-main/backend/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
transformers==4.35.2
torch>=2.2.0
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.4
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
textblob==0.17.1
slowapi==0.1.9
redis==5.0.1
email-validator